moved query_result_type into its own file
